// src/config/env.d.ts
export declare const config: {
    port: string | number;
    supabase: {
        url: string;
        anonKey: string;
        serviceRoleKey: string;
    };
    resend: {
        apiKey: string;
        fromEmail: string;
    };
    vapi: {
        secret: string;
    };
    google: {
        clientId: string;
        clientSecret: string;
        redirectUri: string;
    };
};
//# sourceMappingURL=env.d.ts.map